+++ src/main/java/plugily/projects/villagedefense/events/spectator/SpectatorSettingsMenu.java
@@ -19,7 +19,6 @@
 package plugily.projects.villagedefense.events.spectator;
 
 import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
@@ -31,19 +30,22 @@ import org.bukkit.potion.PotionEffect;
 import org.bukkit.potion.PotionEffectType;
 import pl.plajerlair.commonsbox.minecraft.compat.XMaterial;
 import pl.plajerlair.commonsbox.minecraft.item.ItemBuilder;
+import plugily.projects.villagedefense.Main;
+import plugily.projects.villagedefense.handlers.ChatManager;
 
-@Deprecated //api subject to merge
 public class SpectatorSettingsMenu implements Listener {
 
+  private Main plugin;
   private final String inventoryName;
   private final String speedOptionName;
   private Inventory inv;
 
   public SpectatorSettingsMenu(JavaPlugin plugin, String inventoryName, String speedOptionName) {
+    this.plugin = (Main) plugin;
     this.inventoryName = inventoryName;
     this.speedOptionName = speedOptionName;
     plugin.getServer().getPluginManager().registerEvents(this, plugin);
-    initInventory();
+    inv = initInventory();
   }
 
   public void openSpectatorSettingsMenu(Player player) {
@@ -52,7 +54,7 @@ public class SpectatorSettingsMenu implements Listener {
 
   @EventHandler
   public void onSpectatorMenuClick(InventoryClickEvent e) {
-    if (e.getInventory() == null || !e.getView().getTitle().equals(color(inventoryName))) {
+    if (e.getInventory() == null || !e.getView().getTitle().equals(plugin.getChatManager().colorRawMessage(inventoryName))) {
       return;
     }
     if (e.getCurrentItem() == null || !e.getCurrentItem().hasItemMeta()) {
@@ -83,23 +85,20 @@ public class SpectatorSettingsMenu implements Listener {
     }
   }
 
-  private void initInventory() {
+  private Inventory initInventory() {
     Inventory inv = Bukkit.createInventory(null, 9 * 3, inventoryName);
+    ChatManager chatManager = plugin.getChatManager();
     inv.setItem(11, new ItemBuilder(Material.LEATHER_BOOTS)
-        .name(color(speedOptionName + " I")).build());
+        .name(chatManager.colorRawMessage(speedOptionName + " I")).build());
     inv.setItem(12, new ItemBuilder(Material.CHAINMAIL_BOOTS)
-        .name(color(speedOptionName + " II")).build());
+        .name(chatManager.colorRawMessage(speedOptionName + " II")).build());
     inv.setItem(13, new ItemBuilder(Material.IRON_BOOTS)
-        .name(color(speedOptionName + " III")).build());
+        .name(chatManager.colorRawMessage(speedOptionName + " III")).build());
     inv.setItem(14, new ItemBuilder(XMaterial.GOLDEN_BOOTS.parseItem())
-        .name(color(speedOptionName + " IV")).build());
+        .name(chatManager.colorRawMessage(speedOptionName + " IV")).build());
     inv.setItem(15, new ItemBuilder(Material.DIAMOND_BOOTS)
-        .name(color(speedOptionName + " V")).build());
-    this.inv = inv;
-  }
-
-  private String color(String message) {
-    return ChatColor.translateAlternateColorCodes('&', message);
+        .name(chatManager.colorRawMessage(speedOptionName + " V")).build());
+    return inv;
   }
 
 }