+++ src/main/java/plugily/projects/villagedefense/kits/KitRegistry.java
@@ -46,7 +46,7 @@ public class KitRegistry {
   private static final List<Kit> kits = new ArrayList<>();
   private static Kit defaultKit = null;
   private static Main plugin;
-  private static final List<Class> classKitNames = Arrays.asList(LightTankKit.class, ZombieFinderKit.class, ArcherKit.class, PuncherKit.class, HealerKit.class, LooterKit.class, RunnerKit.class,
+  private static final List<Class<?>> classKitNames = Arrays.asList(LightTankKit.class, ZombieFinderKit.class, ArcherKit.class, PuncherKit.class, HealerKit.class, LooterKit.class, RunnerKit.class,
           MediumTankKit.class, WorkerKit.class, GolemFriendKit.class, TerminatorKit.class, HardcoreKit.class, CleanerKit.class, TeleporterKit.class, HeavyTankKit.class, ShotBowKit.class,
           DogFriendKit.class, PremiumHardcoreKit.class, TornadoKit.class, BlockerKit.class, MedicKit.class, NakedKit.class, WizardKit.class);
 
@@ -113,7 +113,7 @@ public class KitRegistry {
   private static void setupGameKits() {
     KnightKit knightkit = new KnightKit();
     FileConfiguration config = ConfigUtils.getConfig(plugin, Constants.Files.KITS.getName());
-    for (Class kitClass : classKitNames) {
+    for (Class<?> kitClass : classKitNames) {
       if (config.getBoolean("Enabled-Game-Kits." + kitClass.getSimpleName().replace("Kit", ""))) {
         try {
           Class.forName(kitClass.getName()).newInstance();